Move Dice from api do Die
